Move RoutingService into foundation-ui
