+++ foundation-ui/routing/index.js
@@ -1,0 +1,3 @@
+import RoutingService from './RoutingService';
+
+export default RoutingService;
